Ef and Sql server

+++ FirstWebApp/Models/DataEntities/Product.cs
@@ -5,8 +5,8 @@
         public required int Id { get; init; }
         public string? Name { get; init; }
         public string? Description { get; init; }
-        public decimal Price { get; init; }
-        public decimal? Amount { get; init;}
+        public double Price { get; init; }
+        public int? Amount { get; init;}
         public string?  Producer  { get; init; }
 
     }
