+++ FirstWebApp/Models/Dtos/ProductDto.cs
@@ -5,8 +5,8 @@ public record ProductDto
     public required int Id { get; init; }
     public string? ProductName { get; init; }
     public string? Description { get; init; }
-    public decimal Price { get; init; }
+    public double Price { get; init; }
 }
 
 
-public record ProductDtoSingleLine(int id, string name, string desc, decimal price);
+public record ProductDtoSingleLine(int id, string name, string desc, double price);
